requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
sqlalchemy==2.0.23
sqlite3
python-dotenv==1.0.0
schedule==1.2.0
gspread==5.12.0
oauth2client==4.1.3
openpyxl==3.1.2
matplotlib==3.8.2
plotly==5.17.0
dash==2.14.2
smtplib
email-validator==2.1.0
python-telegram-bot==20.7
slack-sdk==3.23.0
fake-useragent==1.4.0
requests-html==0.10.0
lxml==4.9.3
webdriver-manager==4.0.1
python-dateutil==2.8.2 